TEMP: ARM: DRA7xx: crossbar: Add support for crossbar
DRA7xx has a large number of interrupts/dma requests to service
the needs of its many peripherals and subsystems. All of the
requests lines from the subsystems are not needed at the same
time, so they have to be muxed to the controllers appropriately.
In such places a interrupt/dma controllers are preceded by an
IRQ/DMA CROSSBAR that provides flexibility in muxing the device
requests to the controller inputs.

This patch adds support for this IRQ/DMA CROSSBAR. This also configures
some of the reserved MPU IRQs.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
[s-anna@ti.com: adjust mailbox irqs]
Signed-off-by: Suman Anna <s-anna@ti.com>

Change-Id: Iaf258fbd3e099f79fedac297e3d4fbffcffc6dea