arm: omap5: dra7xx: Rework bootcmd to handle two MMC devs
omap5_uevm and dra7xx_evm both can boot from either the MMC card
or eMMC chip on board. We should try both interfaces.

This modification also allows a graceful fallback if
a device exists but boot images are not present on it.

This is adopted from am335x_evm patch.

Change-Id: I072ef56be324c9d343198a05fa26af66e3a9830e
Signed-off-by: alaganraj <alaganraj.s@ti.com>
[Resolved merge conflicts and added change id for gerrit]
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>