u-boot: Changes for loading ipu1 with Sysbios image
Added changes to load ipu1 from cmd_booti. The ELF binary
is placed in a separate ipu partition and read in.
The file thus read is elf parsed and loaded into memory.
The various ipu1 peripherals are initialized and the ipu1 is
brought out of reset, made to execute the loaded image.

Change-Id: Iaad9fdfdaa7373e68900b9049e6eea5d8e69c2e3
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>
Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>
Signed-off-by: Sundar Raman <sunds@ti.com>