DRA7XX: Add support to access NOR flash from MMC boot mode
This patch adds supports to access the NOR flash from the
MMC boot mode. This enables to flash the u-boot.bin to
NOR flash.

To access NOR flash in MMC boot mode use dra7xx_evm_nor
configuration file and build the u-boot.

Note:- The NOR flash cannot be accessed from QSPI as the GPMC
lines are pinmuxed with NOR.

Change-Id: Ie2a97c3bcf51e2ba6d60201f3b979004d1b07a2f
Signed-off-by: Lad, Prabhakar <prabhakar.lad@ti.com>